fix: add short-circuit behavior to compare-with-uploaded.luau (#6)
`compare-with-uploaded.luau` currently tells you every single descendant
that was removed if an ancestor was removed. This pull short-circuits on
descendant added/removed so that only the highest-level ancestor is
marked as removed.

We probably don't need to do this for additions because adding multiple
layers of objects in one PR should still be flagged in CI, as all of
them are being explicitly created. This is in contrast to deletion,
where deleting a top-level object deletes the descendants by proxy.